+++ resources/views/components/dropdown/index.blade.php
@@ -7,15 +7,16 @@
     'width' => null,
 ])
 
-<div {{ $attributes->class(['aui-dropdown']) }} x-data="auiDropdown()" x-on:click.outside="hide()" x-on:keydown.escape.window="hide()">
+<div {{ $attributes->class(['aui-dropdown']) }} x-data="auiDropdown({ align: @js($align) })" x-on:click.outside="hide()" x-on:keydown.escape.window="hide()">
     @isset($trigger)
-        <div x-on:click="toggle()">{{ $trigger }}</div>
+        <div x-ref="trigger" x-on:click="toggle()">{{ $trigger }}</div>
     @else
         <x-avian-ui::button
             :variant="$variant"
             :size="$size"
             :icon="$icon"
             icon-right="fas fa-chevron-down"
+            x-ref="trigger"
             x-on:click="toggle()"
             x-bind:aria-expanded="open"
             aria-haspopup="true"
@@ -25,7 +26,9 @@
     <div
         x-cloak
         x-show="open"
+        x-ref="menu"
         x-on:click="select()"
+        x-bind:style="{ top: top + 'px', left: left + 'px' }"
         @class(['aui-dropdown-menu', 'aui-dropdown-menu-right' => $align === 'right'])
         @if (filled($width)) style="min-width: {{ $width }}" @endif
     >